pkg/util: Mv net devices requirement logic under SIG network
Currently, `NeedVirtioNetDevice` and `NeedTunDevice` are located under
a general purpose `pkg/util`.
Move the above functions under SIG-network's ownership.
