+++ pkg/util/BUILD.bazel
@@ -11,7 +11,6 @@ go_library(
     importpath = "kubevirt.io/kubevirt/pkg/util",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/network/vmispec:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubevirt/scheme:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
